Re-wrote to have a single request method
Rather than having multiple request methods with the same code, this
Class was re-written to have a single request method to utilize
requests.request method. The auth was also updated to utilize
request_oauthlibs OAuth1 auth process. This allows us to not have to
maintain our own version of OAuth standards and can rely on the
requests_oauthlib version.